requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
PyMuPDF
pdfplumber
python-docx
langchain
google-generativeai
streamlit
requests
python-dotenv
PyMuPDF
docx2txt
textract
streamlit
pandas